Load in onclick callbacks. Supports BuySell callback.

+++ bNpcStore/src/uk/codingbadgers/bnpcstore/trait/gui/callbacks/GuiBuySellCallback.java
@@ -7,7 +7,6 @@
 package uk.codingbadgers.bnpcstore.trait.gui.callbacks;
 
 import org.bukkit.Material;
-import org.bukkit.enchantments.Enchantment;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.ItemStack;
@@ -20,7 +19,7 @@ public class GuiBuySellCallback implements GuiCallback {
     private final GuiInventory m_previousMenu;
     private final GuiInventorySubMenu m_buySellMenu;
     
-    public GuiBuySellCallback(GuiInventory previousMenu, String name, Material item, int buyPrice, int sellPrice) {
+    public GuiBuySellCallback(GuiInventory previousMenu, String name, Material item, double buyPrice, double sellPrice) {
         m_previousMenu = previousMenu;
         m_buySellMenu = new GuiInventorySubMenu(m_previousMenu, "Buy/Sell " + name, 5);
